saup/requirements.txt
openai>=1.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
torch>=2.0.0
rich>=13.0.0
wikipedia>=1.4.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
